As per the suggestion moved cocktailPage to container folder
