+++ src/components/app/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import CocktailListPage from '../../pages/cocktailListPage/CocktailListPage';
-import CocktailPage from '../../pages/cocktailPage/CocktailPage'; // eslint-disable-line import/no-named-as-default
+import CocktailPage from '../../container/cocktailPage/CocktailPage'; // eslint-disable-line import/no-named-as-default
 
 import './App.css';
